[FEAT] *: FPS Improvements from 2 min to 10 ms
This PR introduces the batch intersect. This method calculates all
triangles intersection for a take. However, it is not possible do
phong or ray tracing anymore, only ray casting, if you use the
rayTracing=False flag on the camera.take() method.

It is possible now to render 80.000 triangles in incredible 0.10
millisseconds, possibiliting the creation of various applications
3D in real time. So, if you want use the ray tracing and phong, it
will possible just for take a photo, not make a real time application.

The context for this PR is entirely related to CPU process, not GPU.
For GPU, more improvements are comming.

It is the start of an incredible history of this engine. I am so happy.